+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 **Discord**: Join the [Discord Server](https://discord.gg/zKWCkwbnn3) for discussion, questions, or requests
 
 # Change Log
+## 1.10.3 - Bug fixes
+![](https://img.shields.io/badge/release%20date-April%2021%2C%202026-blue)
+![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.10.3/module.zip)
+
+- Fixed bug that prevented search from working for some users
+- Dramatically sped up the time to expand/collapse nodes in settings with lots of entries
+
 ## 1.10.2 - Support for Calendaria v1.0
 ![](https://img.shields.io/badge/release%20date-April%205%2C%202026-blue)
 ![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.10.2/module.zip)